RSC: Remove unused code, and move things around to improve code organization
