--- packages/vite/src/waku-lib/rsc-utils.ts
@@ -1,84 +1,0 @@
-// TODO (RSC) Take ownership of this file and move it out ouf the waku-lib folder
-import { Buffer } from 'node:buffer'
-import { Transform } from 'node:stream'
-
-export const codeToInject = `
-    globalThis.__rw_module_cache__ = new Map();
-
-    globalThis.__webpack_chunk_load__ = (id) => {
-      return import(id).then((m) => globalThis.__rw_module_cache__.set(id, m))
-    };
-
-    globalThis.__webpack_require__ = (id) => {
-      return globalThis.__rw_module_cache__.get(id)
-    };\n  `
-
-export const generatePrefetchCode = (
-  basePath: string,
-  entryItemsIterable: Iterable<readonly [rscId: string, props: unknown]>,
-  moduleIds: Iterable<string>
-) => {
-  const entryItems = Array.from(entryItemsIterable)
-  let code = ''
-  if (entryItems.length) {
-    const rscIds = Array.from(new Set(entryItems.map(([rscId]) => rscId)))
-    code += `
-globalThis.__WAKU_PREFETCHED__ = {
-${rscIds
-  .map((rscId) => {
-    const value =
-      '{' +
-      entryItems
-        .flatMap(([id, props]) => {
-          if (id !== rscId) {
-            return []
-          }
-          // FIXME we blindly expect JSON.stringify usage is deterministic
-          const serializedProps = JSON.stringify(props)
-          const searchParams = new URLSearchParams()
-          searchParams.set('props', serializedProps)
-          return [
-            `'${serializedProps}': fetch('${basePath}${rscId}/${searchParams}')`,
-          ]
-        })
-        .join(',') +
-      '}'
-    return `  '${rscId}': ${value}`
-  })
-  .join(',\n')}
-};`
-  }
-  for (const moduleId of moduleIds) {
-    code += `
-import('${moduleId}');`
-  }
-  return code
-}
-
-// HACK Patching stream is very fragile.
-export const transformRsfId = (prefixToRemove: string) => {
-  // Should be something like /home/runner/work/redwood/test-project-rsa
-  console.log('prefixToRemove', prefixToRemove)
-
-  return new Transform({
-    transform(chunk, encoding, callback) {
-      if (encoding !== ('buffer' as any)) {
-        throw new Error('Unknown encoding')
-      }
-      const data = chunk.toString()
-      const lines = data.split('\n')
-      console.log('lines', lines)
-      let changed = false
-      for (let i = 0; i < lines.length; ++i) {
-        const match = lines[i].match(
-          new RegExp(`^([0-9]+):{"id":"${prefixToRemove}(.*?)"(.*)$`)
-        )
-        if (match) {
-          lines[i] = `${match[1]}:{"id":"${match[2]}"${match[3]}`
-          changed = true
-        }
-      }
-      callback(null, changed ? Buffer.from(lines.join('\n')) : chunk)
-    },
-  })
-}
